Move victory points method
