+++ tableau.rb
@@ -5,4 +5,8 @@ class Tableau
   def initialize(cards)
     @cards = cards
   end
+
+  def victory_points
+    @cards.reduce(0){ |acc, ite| acc + ite.victory_points }
+  end
 end
